src/aws-dataprocessing-mcp-server/agentcore-deploy/requirements.txt
awslabs.aws-dataprocessing-mcp-server
mcp
fastmcp
pydantic
httpx
botocore
loguru
boto3
requests
pyyaml
cachetools
